Move balanceUnlocked action to dpos
